package restructure for clean install + minor fixes
