+++ src/dlomix/data/charge_state.py
@@ -1,13 +1,13 @@
 from typing import Callable, Dict, List, Optional, Union
 
-from datasets import Dataset
-
 from ..constants import ALPHABET_UNMOD
 from .dataset import PeptideDataset
 from .dataset_utils import EncodingScheme
 
 
-class FragmentIonIntensityDataset(PeptideDataset):
+class ChargeStateDataset(PeptideDataset):
+    DEFAULT_SPLIT_NAMES = ["train", "val", "test"]
+
     def __init__(
         self,
         data_source: Optional[Union[str, List]] = None,
@@ -15,15 +15,12 @@ class FragmentIonIntensityDataset(PeptideDataset):
         test_data_source: Optional[Union[str, List]] = None,
         data_format: str = "parquet",
         sequence_column: str = "modified_sequence",
-        label_column: str = "intensities_raw",
+        label_column: str = "most_abundant_charge_by_count",
         val_ratio: float = 0.2,
         max_seq_len: Union[int, str] = 30,
         dataset_type: str = "tf",
-        batch_size: int = 64,
-        model_features: Optional[List[str]] = [
-            "precursor_charge_onehot",
-            "collision_energy_aligned_normed",
-        ],
+        batch_size: int = 256,
+        model_features: Optional[List[str]] = None,
         dataset_columns_to_keep: Optional[List[str]] = None,
         features_to_extract: Optional[List[Union[Callable, str]]] = None,
         pad: bool = True,
