+++ src/dlomix/data/fragment_ion_intensity.py
@@ -1,15 +1,11 @@
 from typing import Callable, Dict, List, Optional, Union
 
-from datasets import Dataset
-
 from ..constants import ALPHABET_UNMOD
 from .dataset import PeptideDataset
 from .dataset_utils import EncodingScheme
 
 
-class ChargeStateDataset(PeptideDataset):
-    DEFAULT_SPLIT_NAMES = ["train", "val", "test"]
-
+class FragmentIonIntensityDataset(PeptideDataset):
     def __init__(
         self,
         data_source: Optional[Union[str, List]] = None,
@@ -17,12 +13,15 @@ class ChargeStateDataset(PeptideDataset):
         test_data_source: Optional[Union[str, List]] = None,
         data_format: str = "parquet",
         sequence_column: str = "modified_sequence",
-        label_column: str = "most_abundant_charge_by_count",
+        label_column: str = "intensities_raw",
         val_ratio: float = 0.2,
         max_seq_len: Union[int, str] = 30,
         dataset_type: str = "tf",
-        batch_size: int = 256,
-        model_features: Optional[List[str]] = None,
+        batch_size: int = 64,
+        model_features: Optional[List[str]] = [
+            "precursor_charge_onehot",
+            "collision_energy_aligned_normed",
+        ],
         dataset_columns_to_keep: Optional[List[str]] = None,
         features_to_extract: Optional[List[Union[Callable, str]]] = None,
         pad: bool = True,
