+++ src/dlomix/layers/attention.py
@@ -1,6 +1,6 @@
 import tensorflow as tf
 import tensorflow.keras.backend as K
-from tensorflow.keras import activations, constraints, initializers, regularizers
+from tensorflow.keras import constraints, initializers, regularizers
 
 
 class DecoderAttentionLayer(tf.keras.layers.Layer):
