+++ src/dlomix/losses/__init__.py
@@ -1,3 +1,3 @@
 from .intensity import masked_pearson_correlation_distance, masked_spectral_distance
 
-__all__ = [masked_spectral_distance, masked_pearson_correlation_distance]
+__all__ = ["masked_spectral_distance", "masked_pearson_correlation_distance"]
