+++ src/dlomix/models/base.py
@@ -1,6 +1,6 @@
 import tensorflow as tf
 
-from dlomix.constants import ALPHABET_UNMOD
+from ..constants import ALPHABET_UNMOD
 
 
 class RetentionTimePredictor(tf.keras.Model):
