+++ src/dlomix/pipelines/__init__.py
@@ -1,3 +1,3 @@
 from .pipeline import RetentionTimePipeline
 
-__all__ = [RetentionTimePipeline]
+__all__ = ["RetentionTimePipeline"]
