+++ src/dlomix/reports/IntensityReport.py
